Created model classes
