+++ client/src/pages/Vehicle/Model/Trailer.tsx
@@ -1,0 +1,9 @@
+type Trailer = {
+  id: number;
+  registration: string;
+  makeYear: string;
+  noOfPalletSpaces: number;
+  length: number;
+};
+
+export default Trailer;
