+++ client/src/pages/Vehicle/Model/Truck.tsx
@@ -1,0 +1,9 @@
+type Truck = {
+  id: number;
+  registration: string;
+  makeYear: string;
+  reservoarCapacity: number;
+  horsepower: number;
+};
+
+export default Truck;
